Changes: use REST API endpoints
