+++ config/areas/account.php
@@ -11,7 +11,6 @@ return function () {
 		'dialogs'   => require __DIR__ . '/account/dialogs.php',
 		'drawers'   => require __DIR__ . '/account/drawers.php',
 		'dropdowns' => require __DIR__ . '/account/dropdowns.php',
-		'requests'  => require __DIR__ . '/account/requests.php',
 		'views'     => require __DIR__ . '/account/views.php'
 	];
 };
