+++ config/areas/users.php
@@ -12,7 +12,6 @@ return function ($kirby) {
 		'dialogs'   => require __DIR__ . '/users/dialogs.php',
 		'drawers'   => require __DIR__ . '/users/drawers.php',
 		'dropdowns' => require __DIR__ . '/users/dropdowns.php',
-		'requests'  => require __DIR__ . '/users/requests.php',
 		'searches'  => require __DIR__ . '/users/searches.php',
 		'views'     => require __DIR__ . '/users/views.php'
 	];
