+++ phpunit.xml.dist
@@ -31,6 +31,7 @@
 			<directory suffix=".php">./config/areas/lab</directory>
 			<directory suffix=".php">./config/blocks</directory>
 			<directory suffix=".php">./config/templates</directory>
+			<file>./config/api/routes/changes.php</file>
 			<file>./config/areas/account/buttons.php</file>
 			<file>./config/areas/files/buttons.php</file>
 			<file>./config/areas/languages/buttons.php</file>
@@ -38,10 +39,8 @@
 			<file>./config/areas/users/buttons.php</file>
 			<file>./config/aliases.php</file>
 			<file>./config/setup.php</file>
-			<file>./config/areas/account/requests.php</file>
 			<file>./config/areas/files/requests.php</file>
 			<file>./config/areas/site/requests.php</file>
-			<file>./config/areas/users/requests.php</file>
 		</exclude>
 	</source>
